# Improve performance of From<FixedDecimal> for PluralOperands