Before this change, I got:

```
operands/create/from_fixed_decimal
                        time:   [246.59 ns 249.77 ns 253.38 ns]
Found 24 outliers among 100 measurements (24.00%)
  16 (16.00%) low severe
  1 (1.00%) low mild
  2 (2.00%) high mild
  5 (5.00%) high severe
```

After this change, I get:

```
operands/create/from_fixed_decimal
                        time:   [213.19 ns 214.49 ns 216.01 ns]
                        change: [-17.609% -16.706% -15.829%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  1 (1.00%) low severe
  2 (2.00%) low mild
  5 (5.00%) high mild
  7 (7.00%) high severe
```