@filmil 2020-10-03
Cool maneuver, sffc@.

F

On Fri, Oct 2, 2020 at 10:10 PM Shane F. Carr <notifications@github.com>
wrote:

> Merged #289 <https://github.com/unicode-org/icu4x/pull/289> into master.
>
> —
> You are receiving this because your review was requested.
> Reply to this email directly, view it on GitHub
> <https://github.com/unicode-org/icu4x/pull/289#event-3836232096>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAB4GMC2SPY4PPYZOJ24PTDSI2W4DANCNFSM4SCGGZKA>
> .
>